feat(dict): accept numeric Yomitan term scores (upstream 3448d6d)
Term::score int->double in yomitan_parser.hpp so non-integer Yomitan
term scores parse. Tag::score stays int (upstream unchanged). Zero
on-disk format risk: process_term_bank does not write score.

Also adds native/hoshidicts/UPSTREAM.md baseline marker (Manhhao/
hoshidicts origin/main tip 3448d6d; Hibiki fork local-changes +
deferred batch2/3 inventory). TODO-621 batch1. BUG-350.

+++ native/hoshidicts/hoshidicts_src/json/yomitan_parser.hpp
@@ -19,7 +19,7 @@ struct Term {
   std::string_view reading;
   std::optional<std::string_view> definition_tags;
   std::string_view rules;
-  int score = 0;
+  double score = 0;
   glz::raw_json_view glossary;
   int64_t sequence = 0;
   std::string_view term_tags;
